fix(extensions): unknown extensions shenanigans
Calling `type("UnknownExtensionX", ...).parse_body(data)` would
instanciate a `siotols.extension.UnkownExtension` instead of the
`type`-created class. This is due to the shenanigans done to decorate
the `parse_body` method the enable DEBUG verbosity. The decorated
method's `cls` was the source-file UnknownExtension because it was kept
inside of the decorator local variables.