imp(extensions): expect known enum in responses
As the various responses usually hold a negociated result, that result
must be have been from the request, as we only include types that we
know of inside the requests, it is only natural that the result must not
be unknown too.

This is only a partial validation as there can be more stuff in the
enumeration than what was actually inside of the requests. One must
still make sure that the result found in the response was indeed present
in the request.