ref: parse methods take stream as input
[Big Rewrite] of the many `parse` and `parse_body` methods.

Take a stream as input instead of raw data, to avoid the many `data`
and `SerialIO` conversions.

Use new approaches to parsing lists, to avoid having to manually
manipulate the list length, and to take advantage of the `read_listvar`
and `read_listvar` (renammed from `read_varint`) methods.

The other goal of this rewrite is to make the many parsers more
consistent with one another, to review them and to fix bugs along the
way.

[Big Rewrite]: https://youtu.be/xCGu5Z_vaps