fix(extensions): HelloRetryRequest exts would not register
Since ServerHello and HelloRetryRequest have the same identifier value,
their hash was identic too. This makes it impossible to store both in a
same set. Use a list instead whenever we use HandshakeType_ so it is
possible to register both classes.

We used a set in the first place just because it is the most correct
abstract type to use. Frankly, as we only ever iter on this structure,
we could had use any collections.abc.Iterable structure.