fix(server_hello): echo_legacy_session_id cannot be void
The client must verify that the echo_legacy_session_id found in the
ServerHello is the same as the legacy_session_id sent via the
ClientHello. We do not implement this verification (because TLS1.3
mandates it may be left empty) but other implementations (e.g. cURL)
that are also compatible with prior versions do verify the echo.