imp(states): transcript-hash and HRR shenanigans
Keep a running hash of all the handshakes sent and received. Hold the
original client-hello and server-hello blobs until the cipher-suite is
negociated. We do this to avoid computing many hashes, one for each
different SHA flavor that the client/server is compatible with, and only
keep the flavor found in the negociated cipher-suite. So we had a choice
to either use the CPU or keep stuff in RAM, we decided to keep stuff in
RAM.

Regarding the shenanigans with Hello Retry Request, blame the RFC[^1]:

> Transcript-Hash(ClientHello1, HelloRetryRequest, ... Mn) =
>     Hash(message_hash ||        /* Handshake type */
>          00 00 Hash.length  ||  /* Handshake message length (bytes) */
>          Hash(ClientHello1) ||  /* Hash of ClientHello1 */
>          HelloRetryRequest  || ... || Mn)
>
>  The reason for this construction is to allow the server to do a
>  stateless HelloRetryRequest by storing just the hash of ClientHello1
>  in the cookie, rather than requiring it to export the entire
>  intermediate hash state (see Section 4.2.2).

As implementor ofs TLS, we can safely say that this construction doesn't
help us do a stateless HRR as we shall keep the HRR payload around to
compute the transcript-hash (or keep Hash(Hash(CH)+HRR) as we do in this
work). Hence there is a mandatory state. We cannot offload that state in
a cookie inside the HRR as this would change its hash.

This problem is known as the "How does TLS 1.3 do Stateless Hello Retry
Request if the HRR is also included in the Transcript Hash?" online[^2].

Hopefully we didn't screw the implementation...

[^1]: https://datatracker.ietf.org/doc/html/rfc8446#autoid-45
[^2]: https://security.stackexchange.com/questions/265686/how-does-tls-1-3-do-stateless-hello-retry-request-if-the-hrr-is-also-included