ref: remove debug logging on serialization/parsing
The two Serializable interfaces were coming with a big ugly hack to tune
`_parse` and `_serialize` to print debugging messages. The messages
included the data to parse/serialize along with the expected structure
of the TLS objects.

Those information were very useful at the dawn of siotls when parsing
was all buggy. Now that siotls is compatible with SSLKEYLOGFILE, it is
easy to start wireshark on the side. Parsing is also more stable.

All the `_struct` attributes stay where they are. They serve as
documentation for lazy-me who doesn't want to look the definition of
every content online every time. In case parsing becomes all buggy
again, they will prove handy.