doc: two more links about crypto primitives
The first one is a study made by Daniel J. Bernstein (designer of
Curve25519 and Curve448) and Tanja Lange (chair of the Coding Theory
and Cryptology group at the Eindhoven University) that do the taxomony
of many elliptic curves for cryptography. Along side each curve is a
list of various tests and whether the specific curve pass or fail at
that test. This study consideres that a curve is safe only if it succeed
at passing all tests.

For the curve that matters to us, i.e. the curves standardized to be
used with TLS 1.3, only Curve25519 and Curve448 are considered safe. The
curves from NIST (p256, p384, p521) are all failing at the same 4 tests
of the safecrypto study: rigidity[^1], ladders[^2], completeness[^3],
and indistinguishability from uniform random strings[^4].

The second one is a decade-old report mandated by the European
Commission to the ENISA (European Union Agency for Cybersecurity). I
havn't read the report in full yet but I think it is interresting to put
that link there, for the sake of mentionning that I trust more my own
government than the USA.

[^1]: https://safecurves.cr.yp.to/rigid.html
[^2]: https://safecurves.cr.yp.to/ladder.html
[^3]: https://safecurves.cr.yp.to/complete.html
[^4]: https://safecurves.cr.yp.to/ind.html