imp(config): make configuration immutable
Inside PEP 543 (A Unified TLS API for Python), the authors suggested to
make the Configuration class immutable. While that PEP has since then be
withdrawn, the ideas inside are still interresting.

We hardly ever change the configuration at runtime and in the places we
did (e.g. for HelloRetryRequest) we simply used `dataclasses.replace`
which makes a modified copy of the origin dataclass.