imp: make SerializationError derives BufferError
It proved annoying that SerializationError was deriving ValueError. In
many places we had to explicitly except SerializationError (to just re-
raise the exception) because other ValueErrors had to be cast into a
IllegalParameter error.

To derive the class from BufferError not only simplifies the code in
those places, that base class just seems more appropriate than
ValueError.