imp: store attributes on states instances
Some states need to save a bit of information for the current connection
but bloating the Connection class with many attributes just isn't
beautiful.

Now the attributes set inside states is saved on the state instance
itself if the attribute name didn't match any of the connection object.

We also changed the API for `_move_to_state` to allow some informations
to be passed from one state to another. This is useful for the cookie
that is only needed when sending the second ClientHello (after a HRR) or
for the private keys for key-share.